new entity lenght and ener into dataset file
